+++ lib/net_retrofit_kit.dart
@@ -12,10 +12,10 @@ export 'src/network/http_constant.dart';
 export 'src/network/http_error.dart';
 export 'src/network/http_method.dart';
 export 'src/network/inet_client.dart';
+export 'src/network/istream_net_client.dart';
 export 'src/network/net_options.dart';
 export 'src/network/net_request.dart';
 export 'src/network/net_interceptor.dart';
-export 'src/network/default/default_net_client.dart';
 
 //// ========================== 生成器相关 ==========================
 export 'src/generate/annotations.dart';